nice: introduce GameSession class for managing ICE state
The GameSession class will deal with ICE state and signaling server
interactions. Currently the signaling server stuff is still in the
main file because it's only run when the local player is created,
perhaps it would be better to connect to the server immediately
on Open() (i.e., when the GameSession class is instantiated.)

Each Player gets their own NiceAgent.